+++ controllers/OverviewController.php
@@ -34,39 +34,6 @@ class OverviewController extends BaseController
         return $this->redirect(Url::toOverview($this->contentContainer));
     }
 
-    /**
-     * @return string
-     * @throws \yii\base\Exception
-     */
-    public function actionList()
-    {
-        if (!$this->hasPages()) {
-            return $this->render('no-pages', [
-                'canCreatePage' => $this->canCreatePage(),
-                'createPageUrl' => $this->contentContainer->createUrl('/wiki/page/edit')
-            ]);
-        }
-
-        $pageSize = Yii::$app->getModule('wiki')->pageSize;
-
-        $query = WikiPage::find()
-            ->contentContainer($this->contentContainer)
-            ->orderBy('title ASC');
-
-        $countQuery = clone $query;
-
-        $pagination = new Pagination(['totalCount' => $countQuery->count(), 'pageSize' => $pageSize]);
-        $query->offset($pagination->offset)->limit($pagination->limit);
-
-        return $this->render('list', [
-            'pages' => $query->all(),
-            'pagination' => $pagination,
-            'homePage' => $this->getHomePage(),
-            'contentContainer' => $this->contentContainer,
-            'hasCategories' => $this->hasCategoryPages(),
-        ]);
-    }
-
 
     /**
      * @return OverviewController|string|\yii\console\Response|\yii\web\Response
@@ -74,10 +41,6 @@ class OverviewController extends BaseController
      */
     public function actionListCategories()
     {
-
-        if (!$this->hasCategoryPages()) {
-            return $this->redirect($this->contentContainer->createUrl('/wiki/overview/list'));
-        }
 
         return $this->render('list-categories', [
             'homePage' => $this->getHomePage(),