+++ views/overview/no-pages.php
@@ -5,10 +5,14 @@
  * @license https://www.humhub.com/licences
  */
 
+use humhub\modules\wiki\widgets\WikiContent;
+use humhub\modules\wiki\widgets\WikiMenu;
+use humhub\modules\wiki\helpers\Url;
+use humhub\widgets\Button;
+
 /* @var $this \humhub\components\View */
 /* @var $contentContainer \humhub\modules\content\components\ContentContainerActiveRecord */
 /* @var $canCreatePage boolean */
-/* @var $createPageUrl string */
 
 humhub\modules\wiki\assets\Assets::register($this);
 ?>
@@ -17,31 +21,22 @@ humhub\modules\wiki\assets\Assets::register($this);
     <div class="panel-body">
 
         <div class="row">
-            <div class="col-lg-10 col-md-9 col-sm-9 wiki-content">
+            <?php WikiContent::begin(['cols' => 10])?>
+
                 <div class="text-center wiki-welcome">
                     <h1><?= Yii::t('WikiModule.base', '<strong>Wiki</strong> Module'); ?></h1>
                     <h2><?= Yii::t('WikiModule.base', 'No pages created yet.  So it\'s on you.<br>Create the first page now.'); ?></h2>
                     <?php if ($canCreatePage): ?>
                         <br>
                         <p>
-                            <a href="<?= $createPageUrl; ?>" data-ui-loader
-                               class="btn btn-primary btn-lg"><?php echo Yii::t('WikiModule.base', 'Let\'s go!'); ?></a>
+                            <?= Button::primary( Yii::t('WikiModule.base', 'Let\'s go!'))->link(Url::toWikiCreate($contentContainer))?>
                         </p>
                     <?php endif; ?>
                 </div>
-            </div>
 
-            <div class="col-lg-2 col-md-3 col-sm-3 wiki-menu">
-                <ul class="nav nav-pills nav-stacked">
-                    <?php if ($canCreatePage): ?>
-                        <li>
-                            <a href="<?= $createPageUrl; ?>">
-                                <i class="fa fa-file-text-o new"></i> <?= Yii::t('WikiModule.base', 'New page'); ?>
-                            </a>
-                        </li>
-                    <?php endif; ?>
-                </ul>
-            </div>
+            <?php WikiContent::end() ?>
+
+            <?= WikiMenu::widget(['cols' => 2, 'container' => $contentContainer, 'excludes' => [WikiMenu::LINK_INDEX]]) ?>
 
         </div>
     </div>