+++ views/page/_view_content.php
@@ -5,6 +5,7 @@ use humhub\modules\wiki\helpers\Url;
 
 /* @var $this \humhub\components\View */
 /* @var $page \humhub\modules\wiki\models\WikiPage */
+/* @var $canEdit bool */
 
 
 ?>
@@ -16,7 +17,7 @@ use humhub\modules\wiki\helpers\Url;
 <?php else: ?>
     <br>
     <div class="alert alert-info clearfix">
-        <?= Yii::t('WikiModule.base', 'This page does not contain any content yet.')?><br><br>
-        <?= Button::primary(Yii::t('WikiModule.base', 'Edit page'))->link(Url::toWikiEdit($page))->sm()->icon('fa-pencil-square-o') ?>
+        <?= Yii::t('WikiModule.base', 'This page is empty.')?><br><br>
+        <?= Button::primary(Yii::t('WikiModule.base', 'Edit page'))->link(Url::toWikiEdit($page))->sm()->icon('fa-pencil-square-o')->visible($canEdit) ?>
     </div>
 <?php endif; ?>