+++ views/page/edit.php
@@ -3,8 +3,9 @@
 use humhub\modules\wiki\widgets\WikiEditor;
 use yii\bootstrap\ActiveForm;
 use humhub\modules\wiki\widgets\WikiLinkModal;
-use yii\helpers\Html;
-use humhub\modules\wiki\helpers\Url;
+use humhub\modules\wiki\widgets\WikiMenu;
+use humhub\modules\wiki\widgets\WikiContent;
+use humhub\widgets\Button;
 
 /* @var $this \humhub\components\View */
 /* @var $page \humhub\modules\wiki\models\WikiPage */
@@ -22,89 +23,41 @@ $title = ($page->isNewRecord)
 
 ?>
 <div class="panel panel-default">
-
     <div class="panel-body">
+        <div class="row <?= $page->is_category ? 'wiki-category-page-edit' : '' ?>">
 
-        <div class="row">
-            <div id="wiki-page-edit" class="col-lg-10 col-md-9 col-sm-9 wiki-content">
-
-                <h1><?= $title ?></h1>
+            <?php WikiContent::begin(['title' => $title, 'id' => 'wiki-page-edit', 'cols' => 10]) ?>
 
                 <?php $form = ActiveForm::begin(['enableClientValidation' => false]); ?>
 
-                <?php if ($canAdminister || $page->isNewRecord): ?>
-                    <?= $form->field($page, 'title')->textInput(['placeholder' => Yii::t('WikiModule.views_page_edit', 'New page title')])->label(false); ?>
-                <?php else: ?>
-                    <?= $form->field($page, 'title')->hiddenInput()->label(false); ?>
-                <?php endif; ?>
-
-                <?= $form->field($revision, 'content')->widget(WikiEditor::class)->label(false) ?>
-
-                <script>
-                    $(document).ready(function () {
-                        // Fix MarkdownEditor Url Placeholder, user can also insert wiki page title
-                        $('.linkTarget').attr("placeholder", "<?= Yii::t('WikiModule.views_page_edit', 'Enter a wiki page name or url (e.g. http://example.com)'); ?>");
-                    });
-                </script>
-
-                <?php if ($canAdminister): ?>
-                    <?= $form->field($page, 'is_home')->checkbox(); ?>
-                    <?= $form->field($page, 'admin_only')->checkbox(); ?>
-                    <?= $form->field($page, 'is_category')->checkbox(); ?>
-                    <?php if ($hasCategories): ?>
-                        <?= $form->field($page, 'parent_page_id')->dropDownList($page->getCategoryList()); ?>
-                    <?php endif; ?>
-                <?php endif; ?>
-                <hr>
-                <?= Html::submitButton(Yii::t('WikiModule.views_page_edit', 'Save'), array('class' => 'btn btn-primary', 'data-ui-loader' => true)); ?>
-                <?php ActiveForm::end(); ?>
-            </div>
-
-            <div class="col-lg-2 col-md-3 col-sm-3 wiki-menu">
-                <div class="wiki-menu-fixed">
-                    <?php if (!$page->isNewRecord): ?>
-
-                        <ul class="nav nav-pills nav-stacked">
-                            <?php if ($canAdminister): ?>
-                                <!-- load modal confirm widget -->
-                                <li><?php
-                                    echo \humhub\widgets\ModalConfirm::widget(array(
-                                        'uniqueID' => 'modal_pagedelete_' . $page->id,
-                                        'linkOutput' => 'a',
-                                        'title' => Yii::t('WikiModule.base', '<strong>Confirm</strong> page deleting'),
-                                        'message' => Yii::t('WikiModule.base', 'Do you really want to delete this page?'),
-                                        'buttonTrue' => Yii::t('WikiModule.base', 'Delete'),
-                                        'buttonFalse' => Yii::t('WikiModule.base', 'Cancel'),
-                                        'linkContent' => '<i class="fa fa-trash-o delete"></i> ' . Yii::t('WikiModule.base', 'Delete'),
-                                        'linkHref' => $contentContainer->createUrl('//wiki/page/delete', array('id' => $page->id)),
-                                        'confirmJS' => 'function(jsonResp) { window.location.href = "' . $contentContainer->createUrl('index') . '"; }'
-                                    ));
-                                    ?></li>
-
-                            <?php endif; ?>
-
-                            <li><?= Html::a('<i class="fa fa-reply back"></i> ' . Yii::t('WikiModule.base', 'Cancel'), Url::toWiki($page)); ?></li>
-                            <li class="nav-divider"></li>
-                            <?php if ($homePage !== null) : ?>
-                                <li><?= Html::a('<i class="fa fa-newspaper-o"></i> ' . Yii::t('WikiModule.base', 'Main page'), $contentContainer->createUrl('//wiki/page/index', array())); ?></li>
-                            <?php endif; ?>
-                            <li><?= Html::a('<i class="fa fa-list-alt"></i> ' . Yii::t('WikiModule.base', 'Overview'), $contentContainer->createUrl('//wiki/page/list', array())); ?></li>
-                        </ul>
-
+                    <?php if ($canAdminister || $page->isNewRecord): ?>
+                        <?= $form->field($page, 'title')->textInput(['placeholder' => Yii::t('WikiModule.views_page_edit', 'New page title')])->label(false); ?>
                     <?php else: ?>
-                        <ul class="nav nav-pills nav-stacked">
-                            <li><?= Html::a('<i class="fa fa-reply back"></i> ' . Yii::t('WikiModule.base', 'Cancel'), $contentContainer->createUrl('//wiki/page/list', array('title' => $page->title))); ?></li>
-                            <li class="nav-divider"></li>
-                            <?php if ($homePage !== null) : ?>
-                                <li><?= Html::a('<i class="fa fa-newspaper-o"></i> ' . Yii::t('WikiModule.base', 'Main page'), $contentContainer->createUrl('//wiki/page/index', array())); ?></li>
-                            <?php endif; ?>
-                            <li><?= Html::a('<i class="fa fa-list-alt"></i> ' . Yii::t('WikiModule.base', 'Overview'), $contentContainer->createUrl('//wiki/page/list', array())); ?></li>
-                        </ul>
+                        <?= $form->field($page, 'title')->hiddenInput()->label(false); ?>
                     <?php endif; ?>
-                </div>
-            </div>
-        </div>
 
+                    <?= $form->field($revision, 'content')->widget(WikiEditor::class)->label(false) ?>
+
+                    <?php if ($canAdminister): ?>
+                        <?= $form->field($page, 'is_home')->checkbox(); ?>
+                        <?= $form->field($page, 'admin_only')->checkbox(); ?>
+                        <?= $form->field($page, 'is_category')->checkbox(); ?>
+                        <?php if ($hasCategories): ?>
+                            <?= $form->field($page, 'parent_page_id')->dropDownList($page->getCategoryList()); ?>
+                        <?php endif; ?>
+                    <?php endif; ?>
+
+                    <hr>
+
+                    <?= Button::save()->submit() ?>
+
+                <?php ActiveForm::end(); ?>
+
+            <?php WikiContent::end() ?>
+
+            <?= WikiMenu::widget(['page' => $page, 'edit' => true, 'cols' => 2]) ?>
+
+        </div>
     </div>
 </div>