+++ views/page/view.php
@@ -3,6 +3,8 @@
 use humhub\modules\comment\widgets\CommentLink;
 use humhub\modules\comment\widgets\Comments;
 use humhub\modules\like\widgets\LikeLink;
+use humhub\modules\wiki\widgets\WikiMenu;
+use humhub\modules\wiki\widgets\WikiContent;
 
 
 /* @var $this \humhub\components\View */
@@ -24,35 +26,25 @@ humhub\modules\wiki\assets\Assets::register($this);
     <div class="panel-body">
 
         <div class="row">
-            <div class="col-lg-9 col-md-9 col-sm-9 wiki-content wiki-page-content">
+            <?php WikiContent::begin(['cssClass' => 'wiki-page-content']) ?>
 
-                <?= $this->render('_view_header', ['page' => $page])?>
+            <?= $this->render('_view_header', ['page' => $page]) ?>
 
-                <?= $this->render('_view_content', ['page' => $page]) ?>
+            <?= $this->render('_view_content', ['page' => $page, 'canEdit' => $canEdit, 'content' => $content]) ?>
 
-                <?= $this->render('_view_category_index', ['page' => $page]) ?>
+            <?= $this->render('_view_category_index', ['page' => $page]) ?>
 
-
-                <div class="social-controls">
-                    <?= LikeLink::widget(['object' => $page]); ?>
-                    &middot; <?= CommentLink::widget(['object' => $page]); ?>
-                </div>
-
-                <?= Comments::widget(['object' => $page]); ?>
-
+            <div class="social-controls">
+                <?= LikeLink::widget(['object' => $page]); ?>
+                &middot; <?= CommentLink::widget(['object' => $page]); ?>
             </div>
 
-            <div class="col-lg-3 col-md-3 col-sm-3 wiki-menu">
-                <?= $this->render('_view_menu', [
-                    'page' => $page,
-                    'revision' => $revision,
-                    'homePage' => $homePage,
-                    'canViewHistory' => $canViewHistory,
-                    'canEdit' => $canEdit,
-                    'canAdminister' => $canAdminister,
-                    'canCreatePage' => $canCreatePage,
-                ])?>
-            </div>
+            <?= Comments::widget(['object' => $page]); ?>
+
+            <?php WikiContent::end() ?>
+
+            <?= WikiMenu::widget(['page' => $page, 'revision' => $revision]) ?>
+
         </div>
     </div>
 </div>