+++ widgets/views/categoryListView.php
@@ -3,7 +3,6 @@ use humhub\libs\Html;
 use humhub\modules\wiki\widgets\CategoryListItem;
 
 /* @var $contentContainer \humhub\modules\content\components\ContentContainerActiveRecord */
-/* @var $listUrl string */
 /* @var $categories \humhub\modules\wiki\models\WikiPage[] */
 /* @var $unsortedPages \humhub\modules\wiki\models\WikiPage[] */
 /* @var $options [] */
@@ -18,9 +17,10 @@ use humhub\modules\wiki\widgets\CategoryListItem;
     <?php if(count($unsortedPages)) : ?>
         <?= CategoryListItem::widget([
             'title' => Yii::t('WikiModule.base', 'Pages without category'),
+            'hideTitle' => empty($categories),
             'pages' => $unsortedPages,
-            'contentContainer' => $contentContainer,
-            'url' => $listUrl]) ?>
+            'contentContainer' => $contentContainer
+        ]) ?>
     <?php endif; ?>
 
 <?= Html::endTag('ul') ?>